trying to avoid narrowing conversion

+++ cslibs_math_3d/include/cslibs_math_3d/algorithms/efla_iterator.hpp
@@ -7,7 +7,7 @@
 
 namespace cslibs_math_3d {
 namespace algorithms {
-template <typename Tp = float>
+template <typename Tp = double>
 class EIGEN_ALIGN16 EFLAIterator
 {
 public:
